+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Next Release
 * Added support for creating launch configurations from existing Amazon EC2 instances to the Auto Scaling client
 * Added support for describing Auto Scaling account limits to the Auto Scaling client
 * Added better support for block device mappings to the Amazon AutoScaling client when creating launch configurations
+* [Docs] Updated and added a lot of content in the [User Guide](http://docs.aws.amazon.com/aws-sdk-php/guide/latest/index.html)
 * Fixed a bug where the `KinesisClient::getShardIterator()` method was not working properly
 
 2.5.0 (2013-12-20)